requirements.txt
pandas
numpy
plotly.express
streamlit
matplotlib.pyplot
seaborn